requirements.txt
absl-py==0.7.1      
astor==0.7.1      
attrs == 19.1.0     
backcall ==            0.1.0      
bleach     ==          3.1.0      
cycler       ==        0.10.0     
decorator      ==      4.4.0      
defusedxml       ==    0.6.0      
entrypoints        ==  0.3        
gast   ==              0.2.2      
grpcio   ==            1.20.1     
h5py       ==          2.9.0      
ipykernel    ==        5.1.0      
ipython        ==      7.5.0      
ipython-genutils ==    0.2.0      
ipywidgets         ==  7.4.2      
jedi    ==             0.13.3     
Jinja2    ==           2.10.1     
jsonschema  ==         3.0.1      
jupyter       ==       1.0.0      
jupyter-client  ==     5.2.4      
jupyter-console   ==   6.0.0      
jupyter-core        == 4.4.0      
Keras-Applications  == 1.0.7      
Keras-Preprocessing==  1.0.9      
kiwisolver==           1.1.0      
Markdown    ==         3.1        
MarkupSafe    ==       1.1.1      
matplotlib      ==     3.0.3      
mistune   ==           0.8.4      
mock        ==         3.0.5      
nbconvert     ==       5.5.0      
nbformat        ==     4.4.0      
notebook          ==   5.7.8      
numpy   ==             1.16.3     
pandocfilters ==       1.4.2      
parso           ==     0.4.0      
pexpect      ==        4.7.0      
pickleshare    ==      0.7.5      
Pillow           ==    6.0.0      
pip       ==           19.1.1     
prometheus-client==    0.6.0      
prompt-toolkit     ==  2.0.9      
protobuf    ==         3.7.1      
ptyprocess    ==       0.6.0      
pygame      ==         1.9.5      
Pygments    ==         2.4.0      
pyparsing    ==        2.4.0      
pyrsistent     ==      0.15.1     
python-dateutil  ==    2.8.0      
PyYAML            ==   5.1        
pyzmq       ==         18.0.1     
qtconsole     ==       4.4.4      
scipy           ==     1.2.1      
Send2Trash        ==   1.5.0      
setuptools ==          41.0.0     
six          ==        1.12.0     
tensorboard    ==      1.13.1     
tensorboardX     ==    1.6        
tensorflow         ==  1.13.1     
tensorflow-estimator== 1.13.0     
termcolor    ==        1.1.0      
terminado      ==      0.8.2      
testpath         ==    0.4.2      
torch              ==  1.1.0      
torchvision  ==        0.2.2.post3
tornado        ==      6.0.2      
traitlets        ==    4.3.2      
wcwidth       ==       0.1.7      
webencodings    ==     0.5.1      
Werkzeug          ==   0.15.2     
wheel               == 0.33.1     
widgetsnbextension ==  3.4.2 